Adds empty state to my courses
